chore: comprehensive clippy cleanup with -D warnings for production quality
Apply strict clippy enforcement across 30+ crates (50+ files modified).

Core fixes: astraweave-sdk unsafe blocks, match->matches!, collapsible_if, manual_clamp, field_reassign_with_default, ptr_arg.

Examples: hello_companion RefMut, unified_showcase clamp/unwrap, skinning_demo acronyms, shadow_csm_demo println, biome_showcase struct init, PI constants.

Tests: SDK FFI unsafe blocks (6), astraweave-input flaky test fixes (2).

Verification: clippy --workspace -D warnings PASS, 595+ tests PASS.

+++ tools/aw_editor/src/viewport/widget.rs
@@ -1048,7 +1048,7 @@ impl ViewportWidget {
                         scale = pose.scale,
                         "Captured World start transform"
                     );
-                } else if let Some(entity) = entity_manager.get(selected_id as u64) {
+                } else if let Some(entity) = entity_manager.get(selected_id) {
                     // Fallback to EntityManager
                     self.gizmo_state.start_transform = Some(TransformSnapshot {
                         position: entity.position,
@@ -1332,7 +1332,7 @@ impl ViewportWidget {
             // Transform cancelled - revert to start_transform (NO undo command created)
             if let Some(snapshot) = &self.gizmo_state.start_transform {
                 if let Some(selected_id) = self.selected_entity() {
-                    if let Some(entity) = entity_manager.get_mut(selected_id as u64) {
+                    if let Some(entity) = entity_manager.get_mut(selected_id) {
                         entity.position = snapshot.position;
                         entity.rotation = snapshot.rotation;
                         entity.scale = snapshot.scale;
